Extract css and js to separate files
In a similar fashion to the `pcbdata.json` file,
we can move the styling and scripting to their own files.

This allows us to create a "seam" where we can start to change things.
I.e. how the `index.js` is created (by a human or by a computer)
becomes a separate idea from using the created `index.js`.