Added backend logic for supporting BOM merging by value
Added html for a checkbox and JS logic to interface to the checkbox. The
Checkbox is used for controlling how the BOM is displayed. If checked
the BOM will merge all parts that have the same value and package into
a single entry, otherwise display every part as its own entry.