Eagle ulp exports part attributes
In order to make filtering of parts in the GUI, the part attributes need
to included in the json file. The attributes are included as two entries,
one for the key and one for the value. These are both strings with values
separated by ';'. Processing of the strings is done at the GUI level.

By exporting all of the attributes, this sets up for future development
for dynamically adding part fields. Fo example if I want to display the
manufacture part number, which people use different values for (MPN, MFPN, etc),
they will be able to specify this in the GUI and the attribute strings will
be queried to get the appropriate key and value.