Started implementing additional attribute display
BUG IN THIS COMMIT SEE BELOW

With this the user will be able to specify additional attributes that
they can then display. The program then parses the attribute names and
values. If there is match in the name the corresponding value is displayed.

There is a bug is the user specifies 'bom', the gui oversizes itself