Additional work on the JS backend
There is an issue with using the objects for the parts. JS is using the parts by reference
so they are not able to be processed correctly when creating filtered lists.
To solve this a deep copy function needs to be made. This is then called to
create new modified objects which can then be used in filtered part lists.