Changed how revision and title are displayed
If there is no revision the title will be put in the middle of
the bom. If there is a revision the title goes to the left
and the revision in the middle.

This format was chosen since people read left to right. If a revision
is given then it would make sense under this assumption to have the title
on the far left and the revision in the middle.