Fixed bug with parts with similar names being highlighted
If a parts reference designator was a substring of another part, then
both would be highlighted. This was due to the fact that I have changes
how the highlighted refs are stored. I store them as a comma separated
string, but render expects them as an array of strings.

This commit is a temp patch for fixing the problem. In the future, refs
should become an array.