Finished adding top silkscreen support
All top silkscreen can now be exported from EagleCAD to the intermediate
file.

Implementation of exporting silkscreen for packages comes from
silk.ulp. This ULP provides most of the framework for working with
silkscreen on a layer, but it does not take into account text unless
it is text within a package. To fix this I used the the text exporting
for silkscreen text found in the original iBOM.ulp.