Updated ULP and Python program to render circle pads
Since pads are really square and they become round only when the
roundness parameter is used, I have hard coded a way to display square
and round pads. If the roundness parameter is zero, then the pad is
displayed as a polygon while if it is 100%, the the pad is displayed as
a circle.

No other pad shapes can currently be displayed. Need to come up with a
way to define the shape mathematically in relation to the roundness
parameter before arbitrary shapes with roundness between 0 to 100% can
be displayed.