Added 2 new test boards.
These boards are used to est the different stages of rendering. The BBBW
comes from an example within EagleCAD. This design is very complex and
is being used to test all stages of rendering.

The other brd file was used early on to test that simple silkscreen and
pads could be displayed.