Main GUI python interface
gui.py defines the main function for setting up and running the GUI.
gui.py defines the subclass of the GUI that is generated from
wxFormBuilder.

This file also defines all events that should be processed by the GUI.
The class generated by wxFormBuilder provides empty functions this file
is where the actual implementation of these functions takes place.