Fixed issue with onCheck for placed checkbox.
If "Hide placed parts" is disabled, and a user check the box for placing
a part, the canvas should rerender so that the part is outlined in
green.

This commit address the above problem. Now when a user checks the placed
box, the canvas will be rerendered so that the parts bounding box is
now green.