Example Split.js code for new layouts
The three example files each contain one of the splits tat the system
can be in. These were tested uses JSFiddle.

The key is to not over constrain Split.js as this will cause the split
to fail. It is also important to use the class type split-horizontal
and also set the direction in order to get the splits to behave as they
should.

The examples should be used as a starting update the new layout in iBOM.