Fixed bug with rending traces
Updated trace rending did not filter out traces and instead always
rendered all traces. Commit b335c3d5ee78aae6a45a48c99bfd066895da4522
was the culprit as this removed the check "IsLayerVisible()" which would
call the render function if the trace was to be displayed.

Also made a change to the EagleCAD YLP that now allows add a "Layer"
property to a via. This now makes it so the trace rendering interface
is consistent for all trace elements (lines or vias).