Fixed bug with via rendering
When a via is a polygon shape, the size of the via was being rendered
smaller if a polygon. This was due to the the assumption that the
polygon was inscribed with a circle, where really the polygon
circumscribes the circle.

This commit now calculates the radius that would make the polygon
inscribed. The radius is then used with the original formula in order
to calculate the points of the polygon.